Workign version of zoning
Route discovery packets are distributed through the network, when target
is reached a route discovery response is sent back. This allows for
setting up routes for out of zone communication.

So far there is no route validation or cancelation so if route is broken
it will probably fail to deliver.

Version 7